Rename cpp files to hpp
